--- src/App.js
@@ -1,11 +1,0 @@
-import React, { PureComponent } from 'react';
-
-import Calculator from './components/Calculator';
-
-class App extends PureComponent {
-  render() {
-    return <Calculator />;
-  }
-}
-
-export default App;
