+++ src/components/pages/Quote.jsx
