+++ src/index.css
@@ -1,17 +1,19 @@
 *,
 *::before,
 *::after {
-  padding: 0;
-  margin: 0;
-  box-sizing: inherit;
+	padding: 0;
+	margin: 0;
+	box-sizing: inherit;
 }
-
 html {
-  box-sizing: border-box;
-  font-size: 62.5%;
+	box-sizing: border-box;
+	font-size: 62.5%;
 }
 
 body {
-  font-size: 3rem;
-  background-color: black;
+	font-size: 3rem;
+}
+
+p {
+	font-size: 3.5rem;
 }
